% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TailSeq.R
\name{TailSeq}
\alias{TailSeq}
\title{TailSeq Calculating Trunncation, Extension, and Tail of RNA Molecules}
\usage{
TailSeq(file_path, primerSeq, adaptorSeq, distPrimerFromEoG,
  tailNucleotide, tailThreshold)
}
\arguments{
\item{file_path, }{primerSeq, adaptorSeq, distPrimerFromEoG, tailNucleotide, tailThreshold}
}
\value{
dat_normal
}
\description{
Current packages that take care of sequence-based RNA or DNA sequence although very effective in many of applications, namely Biostrings; in our experience, they cannot capture the specific nuisances 3?\200\231 sequence analysis up to the details we had in our mind. To address this issue, we developed an R package called TailSeq. Please refer to the package manual for details.
}
